Move OneSignal directory to top level
