+++ src/OneSignal/User.ts
